add track filters + refined data module structure
